+++ src/modules/rfps/subcomponents/index.ts
@@ -1,21 +1,1 @@
-import { connect } from 'react-redux';
-import RfpForm from './component';
-
-interface StateProps {
-  user: any;
-}
-
-interface DispatchProps {
-}
-
-const mapStateToProps = (state: any, ownProp?: any): StateProps => ({
-  user: state.user
-});
-
-const mapDispatchToProps = (dispatch: any): DispatchProps => ({
-});
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(RfpForm);
+export * from './nav_tabs';
