+++ src/modules/rfps_form/index.ts
@@ -1,0 +1,21 @@
+import { connect } from 'react-redux';
+import RfpForm from './component';
+
+interface StateProps {
+  user: any;
+}
+
+interface DispatchProps {
+}
+
+const mapStateToProps = (state: any, ownProp?: any): StateProps => ({
+  user: state.user
+});
+
+const mapDispatchToProps = (dispatch: any): DispatchProps => ({
+});
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(RfpForm);
